Add return code when exit
In order to work with makefile gracefully, we need a return code to
detect the situation that the student does not get full score and does
not open the verbose option to see the diff result. Therefore, if this
situation happen, the return code will be 1 which can trigger the
command behind "||". There can be a message output to remind user using
the argument to get the diff result.